refactor: align anychart adapter with adapters/ convention
