+++ src/adapters/anychart/types.ts
