--- vite.anychart.config.ts
@@ -1,45 +1,0 @@
-import path from 'node:path';
-import { defineConfig } from 'vite';
-import dts from 'vite-plugin-dts';
-
-export default defineConfig({
-  plugins: [
-    dts({
-      tsconfigPath: './tsconfig.build.json',
-      rollupTypes: true,
-      insertTypesEntry: false,
-    }),
-  ],
-  build: {
-    lib: {
-      entry: path.resolve(__dirname, 'src/anychart-entry.ts'),
-      formats: ['es'],
-      fileName: () => 'anychart.mjs',
-    },
-    sourcemap: true,
-    outDir: 'dist',
-    emptyOutDir: false,
-    rollupOptions: {
-      onwarn(warning, warn) {
-        if (warning.code === 'MODULE_LEVEL_DIRECTIVE') {
-          return;
-        }
-        warn(warning);
-      },
-    },
-  },
-  define: {
-    'process.env.NODE_ENV': JSON.stringify('production'),
-  },
-  resolve: {
-    alias: {
-      '@command': path.resolve(__dirname, 'src/command'),
-      '@model': path.resolve(__dirname, 'src/model'),
-      '@state': path.resolve(__dirname, 'src/state'),
-      '@service': path.resolve(__dirname, 'src/service'),
-      '@type': path.resolve(__dirname, 'src/type'),
-      '@ui': path.resolve(__dirname, 'src/ui'),
-      '@util': path.resolve(__dirname, 'src/util'),
-    },
-  },
-});
